Riak Elixir Client v1.1.6 Riak.Bucket

Functions

get(bucket)

get(pid, bucket)

Get bucket properties.

keys(bucket)

keys(pid, bucket)

keys(pid, type, bucket)

List all keys in a bucket. This is a potentially expensive operation and should not be used in production.

keys(pid, type, bucket, timeout)

keys!(bucket)

keys!(pid, bucket)

keys!(pid, type, bucket)

List all keys in a bucket. Return only buckets. This is a potentially expensive operation and should not be used in production.

keys!(pid, type, bucket, timeout)

list()

list(pid)

List all buckets on the server with or without server-side timeout. This is a potentially expensive operation and should not be used in production.

list(pid, timeout)

list!()

list!(pid)

List all buckets on the server with or without server-side timeout. Return only buckets. This is a potentially expensive operation and should not be used in production.

list!(pid, timeout)

possible_props()

Used to “register” the atoms used by the protobuf interface, not for public use.

put(bucket, props)

put(pid, bucket, props)

Set bucket properties.

put(pid, bucket, type, props)

reset(bucket)

reset(pid, bucket)

Reset bucket properties back to the defaults.