Riak.CRDT.Counter

Counter data-type on Riak 2.0.

Functions

decrement(counter, amount \\ 1)

Decrement a counter on the amount defaulting in 1

increment(counter, amount \\ 1)

Increment a counter on the amount defaulting in 1

new()

Create a new counter

value(counter)

Get the original value as a number