Riak.CRDT.Map

Encapsulates Riak maps

Functions

delete(map, key)

Delete a key from the map

get(map, key_type, key)

Fetch the value associated to key with the key_type on map

has_key?(map, key)

Test if the key is contained in the map

keys(map)

List all keys of the map

new()

Creates a new map

put(map, key, value)

Update the key on the map by passing the value The value can be any other CRDT

size(map)

Get the map size

update(map, key_type, key, fun)

Update the key on the map by passing the function fun to update the value based on the current value (if exists) as argument The key_type must be :register, :map, :set, :flag or :counter

value(map)

Get the original value of the map