Riak Elixir Client v1.1.5 Riak.Connection

Functions

start(host \\ '127.0.0.1', port \\ 8087, args \\ [])

Create a process to talk with the riak server on host:port.

start_link(host \\ '127.0.0.1', port \\ 8087, args \\ [])

Create a linked process to talk with the riak server on host:port.