Riak Elixir Client v1.1.5 Riak.Object

The Data wrapper makes it convenient to work with Riak data in Elixir

Functions

%Riak.Object{} (struct)

Struct representing a Riak Object. Attributes:

type: String; Bucket Type with a unique name within the cluster namespace

bucket: String; Bucket with a unique name within the bucket type namespace

key: String; Not required; Key with a unique name within the bucket namespace

data: Any; Value to be stored under the key

metadata: Orddict; User specified metadata

vclock: String; Dotted Version Vector / Causal Context for object

content_type: String; Content Type for object

create(args \\ [bucket: "default"])

delete_all_indexes(obj)

delete_all_links(obj)

delete_all_metadata(obj)

delete_index(obj, arg)

delete_link(obj, tag)

delete_metadata(obj, key)

from_robj(robj)

get_all_indexes(obj)

get_all_links(obj)

get_all_metadata(obj)

get_index(obj, arg)

get_link(obj, tag)

get_metadata(obj, key)

Get all metadata entries

index_id(arg)

put_index(obj, arg, values)

put_link(obj, tag, bucket, key)

put_metadata(obj, arg)

to_robj(obj)